add new Size and Cell types
